Extract DownloadThread to VideoDownloadThread and several methods to DownloadTools
